Allow more dynamic action mailer host configuration
We have some emails that got generated with links pointing back
to sparc.musc.edu because that is where the action mailer default
was pointing.

We set this twice because the first time this file is run
our application.yml won't be loaded yet.

Try to not assume that this setting is defined even though there
are probably other things that will break if it isn't, and allow
for both http and https urls.