Moving exception notification to config
It's failing on rsmt-test and so I can't see an error that
is happening.

It looks like this is the way it's supposed to be configured now?
Maybe because it does the rails include, some old gihub issues
seem to alude that the error I'm seeing can be caused by monkey
patches